exercise 4.21

+++ PART4/blog-list/controllers/blogs.js
@@ -16,10 +16,16 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   // adding new blogs is only possible if a valid token is sent with the HTTP POST request
   console.log('request.token added from tokenExtractor middleware', request.token)
+
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  // if decodedToken is verified, format will be {username, id, iat}
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-  if (!(request.token) || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
   }
 
   const user = await User.findById(decodedToken.id)
@@ -38,8 +44,29 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  console.log('request.token added from tokenExtractor middleware', request.token)
+
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  // if decodedToken is verified, format will be {username, id, iat}
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const creatorId = blog.user.toString()
+
+  if (creatorId === decodedToken.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'A blog can be deleted only by the user who added the blog' })
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
